chore: move ImageContent to a separate module
